removed performance actions from actions.ts
